ngfw-15304 Dns server functionality and Api call changes
