+++ untangle-vue-ui/source/src/ngfw/index.js
@@ -1,7 +1,7 @@
 import NgfwAbout from '../components/settings/system/NgfwAbout.vue'
-import NgfwDns from '../components/settings/network/NgfwDns.vue'
+import Dns from '../components/settings/network/Dns.vue'
 import NgfwDhcp from '../components/settings/network/NgfwDhcp.vue'
 import StaticRoutes from '../components/settings/routing/StaticRoutes.vue'
 import Settings from '../components/settings/system/Settings.vue'
 
-export { NgfwAbout, NgfwDns, NgfwDhcp, StaticRoutes, Settings }
+export { NgfwAbout, Dns, NgfwDhcp, StaticRoutes, Settings }
